Temporary removal of a check
This is so that I could test reading and writing to an NVMe namespace
without having the offset be changed by the read and write functions.
This does bring up a bigger issue: why were these reads and writes let
through on something with a supposed size of 0 bytes? First things first
would be to fix the vfs_traverse function / stat function whatnot so
that when a node is initialized it will report its initial stat.