Graph: broke apart vfs_traverse
The massive vfs_traverse function has been broken apart into many
smaller functions. This makes it such that there is no one extremely
complicated function that can be configured in three hundred different
ways, rather many small functions which need little to no configuration
at all. The new implementation is much easier to change in that is
comprised of a base function (internal_vfs_traverse) that handles the
traversal of a filepath, which itself is abstracted to other function,
and many smaller functions which simply implement a new callback that is
passed to the base function in the event that a node does not exist.