Implemented: mount, unmount, open, read, write
Started work on these functions. They need some more refining, and a
change in how the locks are handled as I am going off an old system from
memory instead of what actually works. This should be an easy enough fix
though. The vfs_create_filepath and vfs_load_filepath functions have
been changed to follow the convention of returning the "upto" pointer
instead of just checking the internally. This should allow for the
chaining of functions (i.e.
vfs_create_filepath(vfs_traverse_filepath(filepath, ...), ...) to create
only the filepath that does not exist in the node graph).