VFS: modifications in response to testing
I have finally decided to test all of this VFS code. There were a few
small issues in getting it off the ground. However it is now working
fine. It is still not fully complete, and using the old SMP test does
break it. However, I have added a mechanism for the graph to infer
drivers, and generally made the interface more compact. I have added
debug messages. Mount paths calculated by the internal traversal
function no longer include '/' at the front of the path. Node
information structures are now mandatory to pass to the vfs_create_node
function. I have made it so that the close function doesn't immediately
delete the node rather it caches it. The file descriptors no walso have
a reference counter to ensure that a close operation does not delete a
descriptor while it is still in use by another operation. Corrected the
use of the "upto" pointers. There have also been major simplifications
to the open function. I have implemented linking and a rudementry
listing function. Resolving links caused an issue if a callback function
produced al ink, so I just added some code to break out of link
resolution if the returned path is NULL ahead of re-iterating. The
component base calculation has also been changed to exclude the
separator character. The read, write, and stat functions now make a
local copy of the file descriptor provided so that link resolution can
be kept inside of the VFS instead of dropped on the drivers. The name
comparsion of children in the internal traverse function has also been
cleaned up so that the length of the strings to be compared is not just
the component length but rather the length of the longest string being
compared. This has been introduced as traversing /dev/nvme0n1 would
result in /dev/nvme0 being returned.