Fixed multiple file creation
I moved the branch lock from the create node function to the individual
callbacks and made it an expection of the create node function to have
the parent node's branch lock held prior to calling the function. This
allows for a final check of the parent node's children, letting the
function not create a node and instead return the existing node.