Cleaned up link handling
In the graph I have checked over the logic to handle links. It should be
fine. The graph is now also capable of detecting broken links. Some
additional checks have been added to vfs.c (read and write functions
specifically) to ensure that they do resolve links. The logic with links
is: if a link exists, and its link pointer is set then that means the
pointer will always be valid since the node pointed to by the link's
link pointer may never be deleted so long that the link exists. This is
because link resolution incrementes the ref_count of the node that it
resolves to. The internal traverse function will always write the node
pointed to by the given filepath to the end node double pointer if it is
specified.