Implemented most VFS functions
The rename and link functions have been corrected. Some minor changes
were made to graph interface to expose the flags to the caller. This
allows for the traversal function to be told to exclude the last
component of the given filepath (curcial for the rename operation). Then
I just went through and checked over how error cases are handled and how
they are triggered. They should be fine for the rename and link
functions. In the open function, if the create flag is specified the
code that will handle that needs to be corrected to pass a non-NULL
NodeInfo structure. The last two functions that need to be implemented
are the node_delete and the get_relative_path functions.