Node deletion, moved functions
Node deletion is now working. Though I need to check over the logic to
ensure that it is not dumb luck (it is passing the core bashing defined
by the proc_test function). I have also moved the get path from nodes
function over to the graph. This way it can be made use of over there.
The last two things are to be able to recursively delete nodes and to be
able to get the path from nodes. The logic of the close operation has
also been changed slightly to temporarily remove the cache to test under
extreme circumstances. As well as just generally made simpler. The
structures for the resources and the VFS should be reconsidered (i.e. is
a prev pointer needed, is a secondary ref_count needed on the resource
structure, or must branch locks be ticket locks). Also I could possibly
revisit the VFS to make it lockless, though that may be more trouble
than it is worth. I am also thinking about implementing a dcache system.