Implementation of hints
Hints are stored as a union with the resource pointer in each node, so
they do not add to the size of each node. This can be done as directory
nodes do not have a resource. This hint pointer is to allow the super
block driver to store a little bit of information in each directory node
to speed up lookups, as if a hint is given then the VFS assumes to give
the parent directory's hint to the super block driver and calculate the
path such that it is relative to the parent node (or just the name of
the component). If a hint is not given, then the path from the
mountpoint to the node will be used.