Remove, add debug, sanitize values
Some debug printfs have been removed from graph.c, and some ARC_DEBUG
macros have been added in vfs.c (on error cases). The open function now
also sanitizes the file descriptor pointer it is given. This is so that
if the caller forgets to initialize it to NULL, then it will be done by
the function.