Directories with resources, structure optimization
Removed hints and replaced them with resources for directories. This
should offer a less hacky implementation of what hints were meant to be.
Optimized the node structure a little bit by making branch locks
mutexes. Fixed some logic with vfs_read and vfs_write to not fail when a
link failed to resolve (just write to the link itself). I also fixed
driver inference and callback functions in graph.h. Also created a
structure to hold the arguments that will be given to callbacks, but I
am not sure how permanent this is.