Fix possibly accessing NULL value
When finishing up the loading of a node into the filesystem in the load
callback, a check is done to see if the previously attained stat should
be copied into the nodes stat member. This, in the event that the create
function failed, would attempt to access a NULL value. This has been
fixed by adding an additional check to ensure that the return of the
create function is non-NULL.