Begin rewrite using abstract graphs
I have begun the rewrite of the filesystem to use the abstract graph
system implemented in K/lib. This should reduce the complexity of the
vfs significantly and has thus far eliminated the graph.c/h files. I
have reworked some of the function signatures to make them more elegant.